Adds logger
